fix(ui): Allow to configure deeply nested `StyleProvider`
