+++ packages/ui/src/components/StyleProvider/index.ts
@@ -1,1 +1,2 @@
 export * from './StyleProvider'
+export * from './types'
