+++ packages/ui/src/index.ts
@@ -3,5 +3,6 @@ export * from './components'
 export * from './deprecated'
 export { NavigationContext } from './Navigation'
 export type { NavigationLinkProps } from './Navigation'
+export * from './config'
 export * from './utils'
 export * from './types'
